Move threadsafe rand.Rand to a package
We'll need this across multiple sub-packages, so it needs its own package.
