+++ trand/rand.go
@@ -18,24 +18,33 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package tchannel
+// Package trand provides a thread-safe random number generator.
+package trand
 
 import (
 	"math/rand"
 	"sync"
+	"time"
 )
 
-// lockedSource allows a random number generator to be used by multiple goroutines concurrently.
-// The code is very similar to math/rand.lockedSource, which is unfortunately not exposed.
+// lockedSource allows a random number generator to be used by multiple goroutines
+// concurrently. The code is very similar to math/rand.lockedSource, which is
+// unfortunately not exposed.
 type lockedSource struct {
 	sync.Mutex
 
 	src rand.Source
 }
 
-// NewRand returns a rand.Rand that is threadsafe.
-func NewRand(seed int64) *rand.Rand {
+// New returns a rand.Rand that is threadsafe.
+func New(seed int64) *rand.Rand {
 	return rand.New(&lockedSource{src: rand.NewSource(seed)})
+}
+
+// NewSeeded returns a rand.Rand that's threadsafe and seeded with the current
+// time.
+func NewSeeded() *rand.Rand {
+	return New(time.Now().UnixNano())
 }
 
 func (r *lockedSource) Int63() (n int64) {
